refactor: move modelAuditCliParser.ts to correct directory
The file was erroneously placed in src/utils/ instead of src/util/ in commit 567988d2f.
This moves it to the correct location alongside other utility files and updates
all import references accordingly.
